Add Op4 sniper rifle, fix warning
Added Opposing Force Sniper Rifle
Fixed compiler warning (signed unsigned mismatch)
Made enginefuncs_t::pfnPlaybackEvent const correct
Made event_api_t::EV_PlaySound const correct

+++ game/server/CWorld.cpp
@@ -205,7 +205,7 @@ void CWorld::Precache( void )
 	// 63 testing
 	LIGHT_STYLE( 63, "a" );
 
-	for( int i = 0; i < gDecalsSize; i++ )
+	for( size_t i = 0; i < gDecalsSize; i++ )
 		gDecals[ i ].index = DECAL_INDEX( gDecals[ i ].name );
 
 	// init the WorldGraph.
